requirements.txt
apiclient==1.0.3
appdirs==1.4.3
boto==2.48.0
certifi==2017.7.27.1
chardet==3.0.4
ffmpy==0.2.2
google-api-python-client==1.6.2
httplib2==0.10.3
idna==2.5
instapush==0.1.2
numpy==1.13.1
oauth2client==4.0.0
packaging==16.8
picamera==1.13
pyasn1==0.3.2
pyasn1-modules==0.0.11
pyparsing==2.2.0
requests==2.18.3
RPi.GPIO==0.6.3
rsa==3.4.2
scipy==0.19.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
